requirements.txt
streamlit
streamlit-authenticator
bcrypt
PyYAML
sqlalchemy